Use semantically correct integer types for counts and line numbers

+++ include/mu/tiny/test/ExpectFailShell.hpp
@@ -39,7 +39,7 @@ public:
       const char* group_name,
       const char* test_name,
       const char* file_name,
-      size_t line_number
+      int_least32_t line_number
   );
   ~ExpectFailShell() override = default;
   ExpectFailShell(const ExpectFailShell&) = delete;
