+++ include/mu/tiny/test/Ordered.hpp
@@ -110,7 +110,7 @@ public:
       const char* group_name,
       const char* test_name,
       const char* file_name,
-      size_t line_number,
+      int_least32_t line_number,
       int level
   ) noexcept;
   virtual ~OrderedInstaller() = default;
