+++ src/test/ExpectFailShell.cpp
@@ -13,7 +13,7 @@ ExpectFailShell::ExpectFailShell(
     const char* group_name,
     const char* test_name,
     const char* file_name,
-    size_t line_number
+    int_least32_t line_number
 )
   : Shell(group_name, test_name, file_name, line_number)
 {
