+++ tests/src/MockFailureReporter.hpp
@@ -5,6 +5,8 @@
 
 #include "mu/tiny/mock.hpp"
 
+#include <stdint.h>
+
 #define CHECK_EXPECTED_MOCK_FAILURE(expectedFailure)                           \
   check_expected_mock_failure_location(expectedFailure, __FILE__, __LINE__)
 #define CHECK_EXPECTED_MOCK_FAILURE_STRING(expectedString)                     \
@@ -42,13 +44,13 @@ void clear_mock_failure();
 void check_expected_mock_failure_string_location(
     const mu::tiny::String& expected_string,
     const char* file,
-    size_t line
+    int_least32_t line
 );
 void check_expected_mock_failure_location(
     const mu::tiny::mock::Failure& expected_failure,
     const char* file,
-    size_t line
+    int_least32_t line
 );
-void check_no_mock_failure_location(const char* file, size_t line);
+void check_no_mock_failure_location(const char* file, int_least32_t line);
 
 #endif
